chore(DSFR): Rename component
