+++ front/src/App.vue
@@ -1,11 +1,11 @@
 <script setup>
-import HelloWorld from './components/HelloWorld.vue'
+import GristContext from './components/GristContext.vue'
 </script>
 
 <template>
   <header>
     <div class="wrapper">
-      <HelloWorld />
+      <GristContext />
     </div>
   </header>
 </template>
